add mint api tests
